0.0.1 rc

+++ vscx-template.sh
@@ -31,27 +31,32 @@ set -eo pipefail
 script_name="${0##*/}"
 script_ver="1.0"
 script_title="Visual Studio Code Extension Template"
-script_options=""
+script_options="f:"
 script_switches="chioy"
 
 function usage() {
     end "$script_title (v$script_ver)
 
-Use: $script_name [-o] create
-     $script_name [-cio] project $(text_underline name) $(text_underline publisher) $(text_underline path)
+Use: $script_name [-fo] create
+     $script_name [-cfio] project $(text_underline name) $(text_underline publisher) $(text_underline path)
 
 Actions:
  create   Create the VSCode extension template files (latest and versioned) in $templates_directory directory
  project  Create a new VSCode extension project from template file
-           - $(text_underline name)      Extension name
-           - $(text_underline publisher) Extension publisher ID
-           - $(text_underline path)      Directory path to create project directory in
+           • $(text_underline name)      Extension name
+           • $(text_underline publisher) Extension publisher ID
+           • $(text_underline path)      Directory path to create project directory in
+
+Options:
+ -f $(text_underline path)  Template file to create (create; default: $templates_directory/vscx-template-latest.zip) or use (project)
+           • With create action, a versioned file is not created
+           • With project action, the file at $(text_underline path) is used instead of the latest template file
 
 Switches:
  -c  Open project in VSCode after creating
  -h  View this help
  -i  Install dependencies after creating project (npm install)
- -o  Overwrite template versioned file (create) or project directory (project)
+ -o  Overwrite template versioned file (create), template file (create with -f:) or project directory (project)
  -y  Confirm yes to run
 "
 }
@@ -93,16 +98,26 @@ function confirm_run() {
     [[ $REPLY =~ ^[Yy]$ ]] || end
 }
 
+create_versioned_file=1
+custom_template_file=0
 install_dependencies=0
 open_after_create=0
 overwrite=0
 pkg_json_file="./package.json"
+
 templates_directory="./templates"
+template_file="$(realpath $templates_directory)/vscx-template-latest.zip"
+
 yes=0
 
 while getopts "${script_options}${script_switches}" OPTION; do
     case "$OPTION" in
     c) open_after_create=1 ;;
+    f)
+        template_file="$OPTARG"
+        create_versioned_file=0
+        custom_template_file=1
+        ;;
     h) usage ;;
     i) install_dependencies=1 ;;
     o) overwrite=1 ;;
@@ -116,8 +131,6 @@ shift $(($OPTIND - 1))
 script_action=$1
 
 [[ "$script_action" =~ ^(create|project)$ ]] || end "Invalid action: $script_action" 1
-
-template_latest_file="$(realpath $templates_directory)/vscx-template-latest.zip"
 
 if [[ "$script_action" == "project" ]]; then
     ext_name=$(echo "$2" | tr '[:upper:]' '[:lower:]' | sed 's/ /-/g')
@@ -135,7 +148,7 @@ if [[ "$script_action" == "project" ]]; then
 
     ext_id="$ext_publisher.$ext_name"
 
-    [[ -f "$template_latest_file" ]] || end "Template file not found: $template_latest_file" 1
+    [[ -f "$template_file" ]] || end "Template file not found: $template_file" 1
 
     ext_path=$(realpath "$ext_path")
     ext_directory="$ext_path/vscode-$ext_name"
@@ -150,19 +163,40 @@ if [[ "$script_action" == "project" ]]; then
     [[ $overwrite -ne 1 ]] || echo "• Overwrite existing directory (if exists)"
     [[ $install_dependencies -ne 1 ]] || echo "• Install NPM dependencies"
     [[ $open_after_create -ne 1 ]] || echo "• Open project in VSCode after creating"
-else
-    [[ -f "$pkg_json_file" ]] || end "File not found: $pkg_json_file" 1
-    pkg_version=$(jq -r '.version' package.json)
-    [[ "$pkg_version" == "" || "$pkg_version" == "null" ]] && end "Failed to get version from $pkg_json_file" 1 || true
+    echo "• Template file: $template_file"
 
-    template_versioned_file="$(realpath $templates_directory)/vscx-template-v$pkg_version.zip"
-    [[ -f "$template_versioned_file" && $overwrite -eq 0 ]] && end "File already exists (use -o to overwrite): $template_versioned_file" 1
+else # create
+    if [[ $create_versioned_file -eq 1 ]]; then
+        [[ -f "$pkg_json_file" ]] || end "File not found: $pkg_json_file" 1
+        pkg_version=$(jq -r '.version' package.json)
+        [[ "$pkg_version" == "" || "$pkg_version" == "null" ]] && end "Failed to get version from $pkg_json_file" 1 || true
+
+        template_versioned_file="$(realpath $templates_directory)/vscx-template-v$pkg_version.zip"
+        [[ -f "$template_versioned_file" && $overwrite -eq 0 ]] && end "File already exists (use -o to overwrite): $template_versioned_file" 1
+    fi
+
+    if [[ $custom_template_file -eq 1 ]]; then
+        if [[ -f "$template_file" ]]; then
+            if [[ $overwrite -eq 1 ]]; then
+                echo "Removing existing file..."
+                rm -f "$template_file"
+            else
+                end "File already exists (use -o to overwrite): $template_file" 1
+            fi
+        fi
+    fi
 
     echo "Create VSCode extension template:"
-    echo "• Version: $pkg_version"
-    echo "• Current file: $template_latest_file"
-    echo "• Versioned file: $template_versioned_file"
-    [[ $overwrite -ne 1 ]] || echo "• Overwrite existing versioned file (if exists)"
+    [[ $create_versioned_file -ne 1 ]] || echo "• Version: $pkg_version"
+    echo "• Template file: $template_file"
+    [[ $create_versioned_file -ne 1 ]] || echo "• Versioned file: $template_versioned_file"
+    if [[ $overwrite -eq 1 ]]; then
+        if [[ $custom_template_file -ne 1 ]]; then
+            echo "• Overwrite existing versioned file (if exists)"
+        else
+            echo "• Overwrite existing template file (if exists)"
+        fi
+    fi
 fi
 
 confirm_run
@@ -181,7 +215,7 @@ if [[ "$script_action" == "project" ]]; then
     mkdir -p "$ext_directory"
 
     echo "Unzipping template files to project directory..."
-    unzip -d "$ext_directory" "$template_latest_file" || end "Failed to unzip file" 1
+    unzip -d "$ext_directory" "$template_file" || end "Failed to unzip file" 1
 
     echo "Updating project files..."
     pushd "$ext_directory" >/dev/null
@@ -194,6 +228,8 @@ if [[ "$script_action" == "project" ]]; then
     sed -i '' "s/change-me/$ext_id/g" src/VSCodeExtension/ExtensionTemplateVSCodeExtensionSettings.ts
 
     popd >/dev/null
+
+    echo "Finished creating project: $ext_directory"
 
     if [[ $install_dependencies -eq 1 ]]; then
         echo "Installing NPM dependencies..."
@@ -211,7 +247,7 @@ if [[ "$script_action" == "project" ]]; then
     fi
 
 elif [[ "$script_action" == "create" ]]; then
-    if [[ -f "$template_versioned_file" ]]; then
+    if [[ $create_versioned_file -eq 1 ]] && [[ -f "$template_versioned_file" ]]; then
         if [[ $overwrite -eq 1 ]]; then
             echo "Removing existing versioned file..."
             rm -f "$template_versioned_file"
@@ -220,13 +256,13 @@ elif [[ "$script_action" == "create" ]]; then
         fi
     fi
 
-    if [[ -f "$template_latest_file" ]]; then
+    if [[ -f "$template_file" && $custom_template_file -ne 1 ]]; then
         echo "Removing existing latest file..."
-        rm -f "$template_latest_file"
+        rm -f "$template_file"
     fi
 
-    echo "Creating latest template file..."
-    zip -rX "$template_latest_file" . \
+    echo "Creating template file..."
+    zip -rX "$template_file" . \
         -x ".DS_Store" \
         -x "*/.DS_Store" \
         -x ".git/*" \
@@ -242,8 +278,13 @@ elif [[ "$script_action" == "create" ]]; then
         -x "releases/*" \
         -x "templates/*" \
         -x "vscx-template.sh" \
+        -x "src/Utilities/*" \
         -x "src/VSCodeExtension/ExtensionTemplateVSCodeExtension.ts" \
-        -x "src/VSCodeExtension/ExtensionTemplateVSCodeExtensionSettings.ts" || end "Failed to create zip file" $?
+        -x "src/VSCodeExtension/ExtensionTemplateVSCodeExtensionSettings.ts" \
+        -x "src/VSCodeExtension/VSCodeApi.ts" ||
+        end "Failed to create zip file" $?
+
+    template_file=$(realpath "$template_file")
 
     echo "Creating temporary template files..."
     sed -n '/^# Debugging/,$p' README.md >"$templates_directory/files/README.md"
@@ -251,10 +292,13 @@ elif [[ "$script_action" == "create" ]]; then
     sed "s/extension-creator/change-me/g" src/VSCodeExtension/ExtensionTemplateVSCodeExtension.ts >"$templates_directory/files/src/VSCodeExtension/ExtensionTemplateVSCodeExtension.ts"
     sed "s/extension-creator/change-me/g" src/VSCodeExtension/ExtensionTemplateVSCodeExtensionSettings.ts >"$templates_directory/files/src/VSCodeExtension/ExtensionTemplateVSCodeExtensionSettings.ts"
 
+    perl -0777 -i -pe 's/\n?[\t ]*\/\/ TEMPLATE-REMOVE-START.*?\/\/ TEMPLATE-REMOVE-END\n?//gs' "$templates_directory/files/src/VSCodeExtension/ExtensionTemplateVSCodeExtension.ts"
+    perl -0777 -i -pe 's/[\t ]*((\/\*\* TEMPLATE-ADD-START)|(TEMPLATE-ADD-END \*\/))\n//gs' "$templates_directory/files/src/VSCodeExtension/ExtensionTemplateVSCodeExtension.ts"
+
     pushd "${templates_directory}/files" >/dev/null
 
     echo "Adding temporary template files to latest template file..."
-    zip -rX "$template_latest_file" * \
+    zip -rX "$template_file" * \
         -x ".DS_Store" \
         -x "*/.DS_Store" || end "Failed to add files to zip" $?
 
@@ -265,6 +309,11 @@ elif [[ "$script_action" == "create" ]]; then
     rm -f "$templates_directory/files/src/VSCodeExtension/ExtensionTemplateVSCodeExtension.ts"
     rm -f "$templates_directory/files/src/VSCodeExtension/ExtensionTemplateVSCodeExtensionSettings.ts"
 
-    echo "Creating versioned template file..."
-    cp "$template_latest_file" "$template_versioned_file"
+    if [[ $create_versioned_file -eq 1 ]]; then
+        echo "Creating versioned template file..."
+        cp "$template_file" "$template_versioned_file"
+    fi
+
+    echo "Template file created: $template_file"
+    [[ $create_versioned_file -ne 1 ]] || echo "Versioned file created: $template_versioned_file"
 fi
